backend/pyproject.toml
[tool.poetry]
name = "robinhood-crypto-bot"
version = "0.1.0"
description = "Cryptocurrency trading bot using official Robinhood API"
authors = ["Devin AI <devin-ai-integration[bot]@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pynacl = "^1.5.0"
pandas = "^2.1.3"
numpy = "^1.25.2"
scikit-learn = "^1.3.2"
xgboost = "^2.0.1"
ta-lib = "^0.4.28"
yfinance = "^0.2.22"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
